backend/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Celery for async tasks
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic for validation
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# AI Providers
anthropic==0.18.1
openai==1.12.0  # For OpenRouter (OpenAI-compatible API)

# Stripe for payments
stripe==8.2.0

# Environment variables
python-dotenv==1.0.1

# CORS - handled by FastAPI middleware
# python-cors not needed

# WebSockets
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
Pillow==10.2.0  # For placeholder image generation
